Added WiFi functionality
Added WiFiSpecific.h for WiFi Shield.
HtmlCode.h has an old reference to Spinner.js.
Added WiFi functionality to GarageDoor.ino. There is a problem that
requires a Garage Door Opener shield modification for WiFi
compatibility. On the GDO shield, I bent pin 7 and added a wire from 9
to 7. So on the WiFi shield, 9 is door 1 and 8 is door 2. New version of
the GDO shield is required.